mixin: init builder.conf for repo commands
Allow the user to manipulate their DNF configuration by initializing the
builder.conf if it does not exist.
