+++ mixin/package.go
@@ -16,7 +16,6 @@ package main
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path/filepath"
 	"runtime"
@@ -79,30 +78,6 @@ func runAddPackage(cmd *cobra.Command, args []string) {
 	fmt.Printf("Added %s package to %s bundle.\n", args[0], args[1])
 }
 
-func setUpMixDir(bundle string, upstreamVer, mixVer int) error {
-	var err error
-	err = os.MkdirAll(filepath.Join(mixWS, "local-rpms"), 755)
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(filepath.Join(mixWS, "builder.conf"),
-		[]byte(builderConf), 0644)
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(filepath.Join(mixWS, "mixversion"),
-		[]byte(fmt.Sprintf("%d", mixVer)), 0644)
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(filepath.Join(mixWS, "mixbundles"),
-		[]byte(fmt.Sprintf("%s\nos-core", bundle)), 0644)
-	if err != nil {
-		return err
-	}
-	return ioutil.WriteFile(filepath.Join(mixWS, "upstreamversion"), []byte(fmt.Sprintf("%d", upstreamVer)), 0644)
-}
-
 func addPackage(pkg, bundle string, build bool) error {
 	var err error
 
@@ -111,11 +86,9 @@ func addPackage(pkg, bundle string, build bool) error {
 		return err
 	}
 	mixVer := ver * 1000
-	if _, err = os.Stat(filepath.Join(mixWS, "builder.conf")); os.IsNotExist(err) {
-		err = setUpMixDir(bundle, ver, mixVer)
-		if err != nil {
-			return err
-		}
+	err = setUpMixDirIfNeeded(bundle, ver, mixVer)
+	if err != nil {
+		return err
 	}
 
 	err = os.Chdir(mixWS)
