+++ mixin/repo.go
@@ -15,7 +15,6 @@
 package main
 
 import (
-	"errors"
 	"fmt"
 	"os"
 
@@ -37,6 +36,7 @@ var addRepoCmd = &cobra.Command{
 	Use:   "add <name> <url>",
 	Short: "Add repo <name> at <url>",
 	Long:  `Add the repo at <url> as a repo from which to pull RPMs for building bundles`,
+	Args:  cobra.ExactArgs(2),
 	Run:   runAddRepo,
 }
 
@@ -44,6 +44,7 @@ var removeRepoCmd = &cobra.Command{
 	Use:   "remove <name>",
 	Short: "Removes repo <name> from the DNF conf used by mixer",
 	Long:  `Remove the repo named <name> from the configured DNF conf used by mixer`,
+	Args:  cobra.ExactArgs(1),
 	Run:   runRemoveRepo,
 }
 
@@ -65,6 +66,7 @@ var setURLRepoCmd = &cobra.Command{
 	Use:   "set-url <name> <url>",
 	Short: "Sets the URL for repo <name> to <url>",
 	Long:  `Sets the URL, for repo <name> to <url>. If <name> does not exist the repo will be added to the configuration.`,
+	Args:  cobra.ExactArgs(2),
 	Run:   runSetURLRepo,
 }
 
@@ -74,6 +76,22 @@ var repoCmds = []*cobra.Command{
 	listReposCmd,
 	initRepoCmd,
 	setURLRepoCmd,
+}
+
+func repoPrep() error {
+	var err error
+
+	ver, err := getCurrentVersion()
+	if err != nil {
+		return err
+	}
+	mixVer := ver * 1000
+	err = setUpMixDirIfNeeded("", ver, mixVer)
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func init() {
@@ -86,8 +104,9 @@ func init() {
 }
 
 func runAddRepo(cmd *cobra.Command, args []string) {
-	if len(args) != 2 {
-		fail(errors.New("add requires exactly two arguments: <repo-name> <repo-url>"))
+	err := repoPrep()
+	if err != nil {
+		fail(err)
 	}
 	b, err := builder.NewFromConfig(config)
 	if err != nil {
@@ -107,8 +126,9 @@ func runAddRepo(cmd *cobra.Command, args []string) {
 }
 
 func runRemoveRepo(cmd *cobra.Command, args []string) {
-	if len(args) != 1 {
-		fail(errors.New("remove requires exactly one argument: <name>"))
+	err := repoPrep()
+	if err != nil {
+		fail(err)
 	}
 	b, err := builder.NewFromConfig(config)
 	if err != nil {
@@ -128,6 +148,10 @@ func runRemoveRepo(cmd *cobra.Command, args []string) {
 }
 
 func runListRepos(cmd *cobra.Command, args []string) {
+	err := repoPrep()
+	if err != nil {
+		fail(err)
+	}
 	b, err := builder.NewFromConfig(config)
 	if err != nil {
 		fail(err)
@@ -145,6 +169,10 @@ func runListRepos(cmd *cobra.Command, args []string) {
 }
 
 func runInitRepo(cmd *cobra.Command, args []string) {
+	err := repoPrep()
+	if err != nil {
+		fail(err)
+	}
 	b, err := builder.NewFromConfig(config)
 	if err != nil {
 		fail(err)
@@ -162,8 +190,9 @@ func runInitRepo(cmd *cobra.Command, args []string) {
 }
 
 func runSetURLRepo(cmd *cobra.Command, args []string) {
-	if len(args) != 2 {
-		fail(errors.New("set-url requires exactly two arguments: <name> <url>"))
+	err := repoPrep()
+	if err != nil {
+		fail(err)
 	}
 
 	b, err := builder.NewFromConfig(config)
